ai_service/requirements.txt
fastapi
uvicorn
sentence-transformers
Pillow
torch
numpy
requests
